Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit c282a51fe94ddc8a32c2fe4eefddfef8a38044b6, reversing
changes made to d3a2379d41a5432526e4b59ab2a7743cb7faa70f.

+++ src/platform/portduino/PortduinoGlue.cpp
@@ -16,7 +16,7 @@ void setBluetoothEnable(bool on)
     // not needed
 }
 
-void cpuDeepSleep(uint32_t msecs)
+void cpuDeepSleep(uint64_t msecs)
 {
     notImplemented("cpuDeepSleep");
 }
