+++ src/platform/rp2040/main-rp2040.cpp
@@ -7,7 +7,7 @@ void setBluetoothEnable(bool on)
     // not needed
 }
 
-void cpuDeepSleep(uint32_t msecs)
+void cpuDeepSleep(uint64_t msecs)
 {
     // not needed
 }
